+++ lib/component/ui/icon_image_provider.dart
@@ -9,12 +9,14 @@ class IconImageProvider extends ImageProvider<IconImageProvider> {
   final double scale;
   final int size;
   final Color color;
+  final Color? bgColor;
 
   IconImageProvider(
     this.icon, {
     this.scale = 1.0,
     this.size = 48,
     this.color = Colors.grey,
+    this.bgColor,
   });
 
   @override
@@ -32,6 +34,12 @@ class IconImageProvider extends ImageProvider<IconImageProvider> {
     final recorder = ui.PictureRecorder();
     final canvas = Canvas(recorder);
     canvas.scale(scale, scale);
+    if (bgColor != null) {
+      // 绘制背景
+      final paint = Paint()..color = bgColor!;
+      final rect = Offset.zero & Size(size.toDouble(), size.toDouble()); // 画布大小
+      canvas.drawOval(rect, paint); // 绘制圆形背景
+    }
     final textPainter = TextPainter(textDirection: TextDirection.rtl);
     textPainter.text = TextSpan(
       text: String.fromCharCode(icon.codePoint),