+++ lib/component/ui/imboy_icon.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+class IMBoyIcon {
+// IMBoyLogo
+  static const IconData imboyLogo = IconData(
+    0xe635,
+    fontFamily: 'IMBoyIcon',
+    matchTextDirection: true,
+  );
+
+  // book 图标
+  static const IconData huawei = IconData(
+    0xe67d,
+    fontFamily: 'IMBoyIcon',
+    matchTextDirection: true,
+  );
+
+  static const IconData sending = IconData(
+    0xea17,
+    fontFamily: 'IMBoyIcon',
+    matchTextDirection: true,
+  );
+
+  static const IconData rightBubble = IconData(
+    0xe613,
+    fontFamily: 'IMBoyIcon',
+    matchTextDirection: true,
+  );
+}