+++ lib/page/group/group_detail/add_member_view.dart
@@ -265,7 +265,7 @@ class AddMemberPageState extends State<AddMemberPage> {
                                 image: DecorationImage(
                                   image: AssetImage(
                                     AssetsService.getImgPath(
-                                        'ic_index_bar_bubble_gray'),
+                                        'index_bar_bubble_gray'),
                                   ),
                                   fit: BoxFit.contain,
                                 ),