+++ lib/page/mine/denylist/denylist_view.dart
@@ -154,7 +154,7 @@ class DenylistPage extends StatelessWidget {
                             image: DecorationImage(
                               image: AssetImage(
                                 AssetsService.getImgPath(
-                                    'ic_index_bar_bubble_gray'),
+                                    'index_bar_bubble_gray'),
                               ),
                               fit: BoxFit.contain,
                             ),